pyproject.toml
[project]
name = "rag-eval-qa-single-doc"
version = "1.0.0"
description = "Generate validated Q/A pairs from individual documents for RAG system evaluation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.0.0",
    "pymupdf>=1.25.0",
    "defusedxml>=0.7.1",
    "langgraph>=1.0.0",
    "langchain>=0.3.0",
    "langchain-openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "jinja2>=3.1.0",
    "beautifulsoup4>=4.12.0",
    "rich>=13.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
]

[project.urls]
Repository = "https://github.com/rsonnen/rag-eval-qa-single-doc"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.8.0",
    "mypy>=1.11.0",
    "bandit[toml]>=1.7.5",
    "pre-commit>=3.5.0",
    "types-defusedxml>=0.7.0",
    "types-PyYAML>=6.0.0",
    "types-beautifulsoup4>=4.12.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"
exclude = [".git", "__pycache__", ".venv", ".pytest_cache", ".mypy_cache", ".ruff_cache"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "ARG", "SIM", "PTH", "RUF", "N", "D", "ANN", "S", "C90", "PLC0415"]
ignore = ["D100", "D104", "D107"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "ANN", "S101"]
# Callback handler must match base class signature exactly
"src/single_doc_generator/tracing.py" = ["ARG002", "ANN401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
mypy_path = "$MYPY_CONFIG_FILE_DIR/src"
namespace_packages = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "pymupdf"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "langgraph.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "langchain_core.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "langchain_openai.*"
ignore_missing_imports = true

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
fail_under = 80
precision = 2
show_missing = true

[tool.pytest.ini_options]
addopts = ["--cov=src", "--cov-report=term-missing", "--cov-report=html", "--cov-fail-under=80", "-v", "-m", "not integration"]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "integration: live LLM tests (cost money, excluded by default). Run with: pytest -m integration",
]

[tool.bandit]
targets = ["src"]
exclude_dirs = ["/tests", "/.venv"]
skips = ["B101"]
